feat: store share counts as scaled integers, so fractions are representable
A broker filling a fixed-amount purchase hands you 2.79 shares. Quantity
was a whole int, so that could not be recorded at all — and rounding it to
3 puts the error into the cost basis on the first entry, where it stays
and compounds through every figure built on one. This is a prerequisite
for recording a savings plan faithfully, not a nicety.

Shares now follow the rule money already follows: an integer on the wire,
scaled, converted once at the UI edge. A millionth of a share is finer
than any broker reports — Schwab and IBKR give four decimals, Robinhood
six — so nothing is truncated coming in. models.Gross is the only place
the scale divides back out; nothing multiplies a quantity by a price
without it, and frontend/src/shares.ts is the only place the frontend
divides.

Invariant 1 changes and CLAUDE.md says so. A cost basis was exact because
it was only ever added to and subtracted from; fractional shares put one
rounding step in front of it, because 2.79 shares at $10.01 is $27.9279
and no representation makes that a whole cent. The rounding happens once,
in Gross, and the integer result is what is stored. A full exit still
lands on exactly 0 — cost x q / q is cost at any scale — so the most
sensitive test of the proportional math is untouched.

The proportional release now computes its product in arbitrary precision.
That was previously defended by a comment saying overflow was not a
practical concern, which was true while quantities were whole shares and
stopped being true the moment they were scaled: a cost basis of NT$92M on
a 100,000-share position passes int64, and a wrapped product is silently
wrong rather than loudly. It runs once per ledger entry on a replay, so
being unconditionally right costs nothing worth having.

The unit change needs a backfill, since AutoMigrate can add a column but
not rewrite one. db.scaleShareCounts multiplies every stored quantity in a
single transaction, and must run exactly once — nothing in the data says
whether it already has, because 100 is a plausible number of whole shares
and a plausible 0.0001 of one. models.SchemaMeta holds the marker that
decides. Verified against a copy of a real book: quantities multiplied,
every cost basis and net amount untouched, and a restart changed nothing.

+++ internal/db/transactions.go
@@ -127,7 +127,7 @@ func (d *DB) ListTransactions(opts ListOptions, filter TransactionFilter) ([]mod
 // instrument are deliberately not editable: changing either would move the entry
 // between positions, which is clearer expressed as a delete plus a re-entry.
 type TransactionUpdate struct {
-	Quantity int
+	Quantity int64
 	Price    int64
 	Fee      int64
 	TradedAt time.Time
